refactor: Completes a first implementation of syncAllOrganizations and syncAllResources for BBMRI
