--- src/main/java/eu/bbmri_eric/negotiator/service/BBMRIDirectoryServiceSynchClient.java
@@ -1,6 +1,0 @@
-package eu.bbmri_eric.negotiator.service;
-
-public interface BBMRIDirectoryServiceSynchClient {
-
-  public void syncAllDiscoveryServiceObjects();
-}
