+++ src/main/java/eu/bbmri_eric/negotiator/service/DiscoveryServiceClient.java
@@ -1,0 +1,10 @@
+package eu.bbmri_eric.negotiator.service;
+
+public interface DiscoveryServiceClient {
+
+  public void syncAllOrganizations();
+
+  public void syncAllResources();
+
+  public void syncAllNetworks();
+}
